requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
pandas==2.0.0
openpyxl==3.1.0
python-dateutil==2.8.2
requests==2.31.0
PyMuPDF==1.24.9
python-docx==1.1.2

# Optional (OCR)
# paddleocr==2.7.3
# paddlepaddle==2.6.1
